+++ clients/download.go
@@ -16,6 +16,7 @@ type PluginType string
 const (
 	PluginTypeSource      PluginType = "source"
 	PluginTypeDestination PluginType = "destination"
+	DefaultDownloadDir               = ".cq"
 )
 
 func DownloadPluginFromGithub(ctx context.Context, localPath string, org string, name string, version string, typ PluginType, writers ...io.Writer) error {
@@ -50,6 +51,7 @@ func DownloadPluginFromGithub(ctx context.Context, localPath string, org string,
 	if org != "cloudquery" {
 		pathInArchive = fmt.Sprintf("cq-%s-%s", typ, name)
 	}
+	pathInArchive = withBinarySuffix(pathInArchive)
 
 	fileInArchive, err := archive.Open(pathInArchive)
 	if err != nil {
@@ -104,4 +106,11 @@ func downloadFile(ctx context.Context, localPath string, url string, writers ...
 	}
 
 	return nil
+}
+
+func withBinarySuffix(path string) string {
+	if runtime.GOOS == "windows" {
+		return path + ".exe"
+	}
+	return path
 }