+++ specs/destination.go
@@ -19,8 +19,13 @@ type Destination struct {
 }
 
 const (
-	WriteModeAppend WriteMode = iota
+	WriteModeOverwriteDeleteStale WriteMode = iota
 	WriteModeOverwrite
+	WriteModeAppend
+)
+
+var (
+	writeModeStrings = []string{"overwrite-delete-stale", "overwrite", "append"}
 )
 
 func (d *Destination) SetDefaults() {
@@ -64,7 +69,7 @@ func (d *Destination) Validate() error {
 }
 
 func (m WriteMode) String() string {
-	return [...]string{"append", "overwrite"}[m]
+	return writeModeStrings[m]
 }
 
 func (m WriteMode) MarshalJSON() ([]byte, error) {
@@ -91,6 +96,8 @@ func WriteModeFromString(s string) (WriteMode, error) {
 		return WriteModeAppend, nil
 	case "overwrite":
 		return WriteModeOverwrite, nil
+	case "overwrite-delete-stale":
+		return WriteModeOverwriteDeleteStale, nil
 	}
 	return 0, fmt.Errorf("invalid write mode: %s", s)
 }